requirements.txt
streamlit
langchain-google-genai
langchain-community
langchain-core
python-dotenv
faiss-cpu
beautifulsoup4